requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiofiles
python-dotenv
groq
pymupdf
pydantic
sqlalchemy
chromadb
sentence-transformers
langchain
langgraph
reportlab
orjson